+++ familias/test_models.py
@@ -1,5 +1,5 @@
 from django.test import TestCase
-from .models import Familia
+from .models import Familia, Oficio
 
 
 class TestFamiliaModel(TestCase):
@@ -20,3 +20,25 @@ class TestFamiliaModel(TestCase):
         of the object.
         """
         self.assertEqual(str(self.familia), 'Molina')
+
+
+class TestOficio(TestCase):
+    """ Unit test suite for testing the Oficio model in .models
+
+    """
+
+    def setUp(self):
+        """ Setup required for all the tests in this suite.
+
+        This setup creates a new Oficio object.
+        """
+        self.nombre = 'Electricista'
+        Oficio.objects.create(nombre=self.nombre)
+
+    def test_str(self):
+        """ Test for the oficio __str__ method.
+
+        This tests that it returns the name of the oficio.
+        """
+        oficio = Oficio.objects.get(nombre=self.nombre)
+        self.assertEqual(str(oficio), self.nombre)
